+++ Google/src/main/java/com/applovin/mediation/adapters/GoogleMediationAdapter.java
@@ -71,6 +71,7 @@ import java.lang.ref.WeakReference;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import androidx.annotation.NonNull;
@@ -473,7 +474,7 @@ public class GoogleMediationAdapter
         if ( isNative )
         {
             NativeAdOptions.Builder nativeAdOptionsBuilder = new NativeAdOptions.Builder();
-            nativeAdOptionsBuilder.setAdChoicesPlacement( NativeAdOptions.ADCHOICES_TOP_RIGHT );
+            nativeAdOptionsBuilder.setAdChoicesPlacement( getAdChoicesPlacement( parameters.getLocalExtraParameters() ) );
             nativeAdOptionsBuilder.setRequestMultipleImages( adFormat == MaxAdFormat.MREC ); // MRECs can handle multiple images via AdMob's media view
 
             NativeAdViewListener nativeAdViewListener = new NativeAdViewListener( parameters, adFormat, activity, listener );
@@ -512,7 +513,7 @@ public class GoogleMediationAdapter
         AdRequest adRequest = createAdRequestWithParameters( isBiddingAd, parameters, activity );
 
         NativeAdOptions.Builder nativeAdOptionsBuilder = new NativeAdOptions.Builder();
-        nativeAdOptionsBuilder.setAdChoicesPlacement( NativeAdOptions.ADCHOICES_TOP_RIGHT );
+        nativeAdOptionsBuilder.setAdChoicesPlacement( getAdChoicesPlacement( parameters.getLocalExtraParameters() ) );
 
         // Medium templates can handle multiple images via AdMob's media view
         String template = BundleUtils.getString( "template", "", parameters.getServerParameters() );
@@ -734,6 +735,22 @@ public class GoogleMediationAdapter
     private boolean isValidNativeAd(final NativeAd nativeAd)
     {
         return nativeAd.getHeadline() != null;
+    }
+
+    private int getAdChoicesPlacement(Map<String, Object> localExtraParams)
+    {
+        // Publishers can set via nativeAdLoader.setLocalExtraParameter( "admob_ad_choices_placement", ADCHOICES_BOTTOM_LEFT );
+        final Object adChoicesPlacementObj = localExtraParams.get( "admob_ad_choices_placement" );
+        return isValidAdChoicesPlacement( adChoicesPlacementObj ) ? (Integer) adChoicesPlacementObj : NativeAdOptions.ADCHOICES_TOP_RIGHT;
+    }
+
+    private boolean isValidAdChoicesPlacement(Object placementObj)
+    {
+        return ( placementObj instanceof Integer ) &&
+                ( (Integer) placementObj == NativeAdOptions.ADCHOICES_TOP_LEFT ||
+                        (Integer) placementObj == NativeAdOptions.ADCHOICES_TOP_RIGHT ||
+                        (Integer) placementObj == NativeAdOptions.ADCHOICES_BOTTOM_LEFT ||
+                        (Integer) placementObj == NativeAdOptions.ADCHOICES_BOTTOM_RIGHT );
     }
 
     //endregion
@@ -1052,7 +1069,14 @@ public class GoogleMediationAdapter
                     }
                     else
                     {
-                        maxNativeAdView = new MaxNativeAdView( maxNativeAd, templateName, activity );
+                        if ( AppLovinSdk.VERSION_CODE >= 11010000 )
+                        {
+                            maxNativeAdView = new MaxNativeAdView( maxNativeAd, templateName, getApplicationContext() );
+                        }
+                        else
+                        {
+                            maxNativeAdView = new MaxNativeAdView( maxNativeAd, templateName, activity );
+                        }
                     }
 
                     nativeAdView = new NativeAdView( activity );