+++ InMobi/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## x.x.x.x
+## 10.0.3.0
+* Certified with InMobi SDK 10.0.3.
+
+## 10.0.2.0
+* Certified with InMobi SDK 10.0.2.
 * Remove `checkExistence()` call from adapter.
 
 ## 10.0.1.0