+++ MyTarget/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.15.0.1
+* Fix a bug where certain native ad assets (icon, MediaView) are obstructed by others.
+
+## 5.15.0.0
+* Certified with MyTarget SDK 5.15.0.
+
 ## 5.14.4.3
 * Add SDK initialization code: `MyTargetManager.initSdk()`.